pr feedback
moved duplicated class to utils
